occt 6.9.1 warning suppression: see following.
    DNaming_Line3DDriver.cxx: enable output to suppress unused warnings
    ViewerTest_ViewerCommands.cxx: reorder constructor for warning suppression
    ViewerTest_RelationCommands.cxx: unused variables
    ViewerTest.cxx: remove unsigned int cast for warning suppression. Why purposefully cast to different types???
    BRepTest_FeatureCommands.cxx: empty statement for warning suppression
    TestTopOpe_CORCommands.cxx: split statement for warning suppression
    BinTools_ShapeSet.cxx: split statement for warning suppression
    Blend_Walking::InternalPerform: adding dummy case for unused enum warning
    BRepFeat_RibSlot.cxx: split stament for warning suppression
    BRepFill_CompatibleWires.cxx: split statement for warning suppression
    TopOpeBRep_FaceEdgeFiller.cxx: split statement for warning suppression
    IntTools_FaceFace.cxx: remove unused variables for warning suppression
    BRepTopAdaptor_FClass2d.cxx: split statements for warning suppression
    Bisector_Inter.cxx: split statements for warning suppression
    GeomFill_GuideTrihedronAC.cxx: split statemenst for warning suppression
    GeomInt_IntSS_1.cxx: remove unused variables causing warnings
    ProjLib_ComputeApprox.cxx: comment out unused global variable for  warning suppression.
    OSD_Path.cxx: added inline keyword to unused functions for warning suppression.

+++ src/ViewerTest/ViewerTest.cxx
@@ -4567,7 +4567,7 @@ Standard_Boolean  ViewerTest::PickObjects(Handle(TColStd_HArray1OfTransient)& ar
   while(NbPickGood<NbToReach && NbPickFail <= MaxPick){
     while(ViewerMainLoop(argccc,argvvv)){}
     Standard_Integer NbStored = TheAISContext()->NbSelected();
-    if((unsigned int ) NbStored != NbPickGood)
+    if(NbStored != NbPickGood)
       NbPickGood= NbStored;
     else
       NbPickFail++;
@@ -4576,7 +4576,7 @@ Standard_Boolean  ViewerTest::PickObjects(Handle(TColStd_HArray1OfTransient)& ar
 
   // step3 get result.
 
-  if((unsigned int ) NbPickFail >= NbToReach) return Standard_False;
+  if(NbPickFail >= NbToReach) return Standard_False;
 
   Standard_Integer i(0);
   for(TheAISContext()->InitSelected();
@@ -4740,7 +4740,7 @@ Standard_Boolean ViewerTest::PickShapes (const TopAbs_ShapeEnum TheType,
   while(NbPickGood<NbToReach && NbPickFail <= MaxPick){
     while(ViewerMainLoop(argccc,argvvv)){}
     Standard_Integer NbStored = TheAISContext()->NbSelected();
-    if((unsigned int ) NbStored != NbPickGood)
+    if(NbStored != NbPickGood)
       NbPickGood= NbStored;
     else
       NbPickFail++;
@@ -4749,7 +4749,7 @@ Standard_Boolean ViewerTest::PickShapes (const TopAbs_ShapeEnum TheType,
 
   // step3 get result.
 
-  if((unsigned int ) NbPickFail >= NbToReach) return Standard_False;
+  if(NbPickFail >= NbToReach) return Standard_False;
 
   Standard_Integer i(0);
   for(TheAISContext()->InitSelected();TheAISContext()->MoreSelected();TheAISContext()->NextSelected()){
